IPEHR-160 small refacorting for filtering query sources
